+++ dash_extensions/_imports_.py
@@ -3,11 +3,15 @@ from .Download import Download
 from .Keyboard import Keyboard
 from .Lottie import Lottie
 from .Monitor import Monitor
+from .Ticker import Ticker
+from .WebSocket import WebSocket
 
 __all__ = [
     "Burger",
     "Download",
     "Keyboard",
     "Lottie",
-    "Monitor"
+    "Monitor",
+    "Ticker",
+    "WebSocket"
 ]